Add common module. Definition of SCRAM Methods
The 'common' module is where most of the SCRAM implementation will
reside, providing functionality to both client and server modules.

Create interface SCRAMMechanism, to define a given mechanism and offer
methods that abstract away how to get a MessageDigest or a Mac for them.

Create enum SCRAMMechanisms that provides some initial implementation of
SCRAM mechanisms, in a simple manner.

Create tests to verify the JVM provides the MessageDigest and Mac
returned by all the enum instances of SCRAMMechanisms.