Remove cross-compilation warning
Compiling with Java > 6 lead to "bootstrap class path not set in
conjunction with -source 1.6" warnings. Since we don't control the build
environment and can effectively set the bootstrap class path to Java6's
rt.jar (that should be done on the release server, though) warnings are
supressed by passing -Xlint:-options to the compiler.