Infrastructure for a generic Sasl Attribute-Value
Both SCRAM and GS2 uses attribute-value pairs where the attribute is a
single char, and the value is a string (value might be null on GS2).

Create CharAttribute and CharAttributeValue interfaces to model the char
attribute and that with the value, respectively. Create a base abstract
implementation that provides some functionality, along with some tests
to try it.