Definition and tests for CharAttributeValueCSV
It is a class with static methods to write and parse CSV
(Comma-Separated Values) of CharAttributeValues. Basically, SCRAM
messages are such a CSV, so this class encapsulates in a generic way
most of the logic required to write and parse such messages.

Writen method definition and several tests to test the implementation.