Implement ClientFirstMessage. Passes tests
Add getters for the fields of the GS2Header. User Optional<String> for
nullable fields return values, to avoid possible NPEs.

Add Preconditions.gt0(int value) method.

Fix a bug in StringWritableCSV that is tested by tests in
ClientFirstMessage, but couldn't be properly exercised in its unit tests
as it is dependent on the full String passed as part of a whole
client-first-message String.