CryptoUtil.nonce() now receives the SecureRandom
Change nonce() method so that it explicitly receives the SecureRandom to
use to generate the nonce. This is important to avoid a SecureRandom
object instantiation each time the method is called. And also allow
callers to provide preferred SecureRandom implementations.

Provide a default method that uses a SecureRandom in a holder pattern.

Write javadoc for the nonce() methods.