Improve scram mechanism selection in ScramClient
There were previously two methods to select the SCRAM mechanism to use
in the client. Both were based on a varags of Strings or a CSV version
of the server's advertised methods.

There are however other ocassions were client method selection should be
explicit. This has been added and previous methods renamed to state the
intent even more clearly.