Add ScramException and refactor exception handling
IllegalArgumentException was heavily used to signal both illegal
arguments (such as calling the public APIs with null or empty Strings
when values were expected) and wrong (semantically speaking) values,
like for example invalid SCRAM messages.

The latter cases have been changed to throw instead the checked
ScramException, which is a child of SaslException, since SCRAM is a SASL
method.